dssd

+++ srcs/packet.c
@@ -44,7 +44,7 @@ void	clearPacket(t_IcmpTargetType *target)
 
 BOOL	bindingPacket(t_IcmpTargetType *targets)
 {
-	targets->packlen = sizeof(struct iphdr) + sizeof(struct icmphdr)
+	targets->packlen = (uint16_t)sizeof(struct iphdr) + sizeof(struct icmphdr)
 		+ targets->currFl.s;
 	targets->packToSend = (char *)malloc(sizeof(char) * targets->packlen);
 	if (!targets->packToSend)
